requirements.txt
ipykernel
langchain
langchain-community
pypdf
sentence-transformers
chromadb
streamlit
google-generativeai
python-dotenv
langchain-huggingface